Fix AMP setup: keep model in float32, autocast handles float16 for forward
GradScaler requires float32 model params — it unscales gradients from
float16 back to float32 before the optimizer step. Having the model in
float16 causes "Attempting to unscale FP16 gradients" error.

+++ train.py
@@ -123,7 +123,7 @@ def build_model(args, device):
         )
         model = Qwen3AttnResForCausalLM(config)
 
-    model = model.to(dtype=torch.float16, device=device)
+    model = model.to(device=device)
     return model
